Move verify implementation from Store to CLI interface
