+++ lib/core/libimagstore/src/store.rs
@@ -323,62 +323,6 @@ impl Store {
         self.configuration.as_ref()
     }
 
-    /// Verify the store.
-    ///
-    /// This function is not intended to be called by normal programs but only by `imag-store`.
-    #[cfg(feature = "verify")]
-    pub fn verify(&self) -> bool {
-        use libimagerror::trace::trace_error_dbg;
-
-        info!("Header | Content length | Path");
-        info!("-------+----------------+-----");
-
-        WalkDir::new(self.location.clone())
-            .into_iter()
-            .all(|res| match res {
-                Ok(dent) => {
-                    if dent.file_type().is_file() {
-                        match self.get(PathBuf::from(dent.path())) {
-                            Ok(Some(fle)) => {
-                                let p           = fle.get_location();
-                                let content_len = fle.get_content().len();
-                                let header      = if fle.get_header().verify().is_ok() {
-                                    "ok"
-                                } else {
-                                    "broken"
-                                };
-
-                                info!("{: >6} | {: >14} | {:?}", header, content_len, p.deref());
-                                true
-                            },
-
-                            Ok(None) => {
-                                info!("{: >6} | {: >14} | {:?}", "?", "couldn't load", dent.path());
-                                true
-                            },
-
-                            Err(e) => {
-                                trace_error_dbg(&e);
-                                if_cfg_panic!("Error verifying: {:?}", e);
-                                debug!("{:?}", e);
-                                false
-                            },
-                        }
-                    } else {
-                        info!("{: >6} | {: >14} | {:?}", "?", "<no file>", dent.path());
-                        true
-                    }
-                },
-
-                Err(e) => {
-                    trace_error_dbg(&e);
-                    if_cfg_panic!("Error verifying: {:?}", e);
-                    debug!("{:?}", e);
-                    false
-                },
-            })
-    }
-
     /// Creates the Entry at the given location (inside the entry)
     ///
     /// # Return value
